fix: move shed back to pkg
