+++ pkg/shed/example_store_test.go
