+++ pkg/shed/field_string_test.go
