+++ pkg/shed/field_struct.go
