+++ pkg/shed/field_struct_test.go
