+++ pkg/shed/schema.go
