+++ pkg/topology/kademlia/internal/metrics/metrics.go
@@ -14,8 +14,8 @@ import (
 	"time"
 
 	"github.com/ethersphere/bee/pkg/p2p"
+	"github.com/ethersphere/bee/pkg/shed"
 	"github.com/ethersphere/bee/pkg/swarm"
-	"github.com/ethersphere/bee/pkg/topology/kademlia/internal/shed"
 	"github.com/syndtr/goleveldb/leveldb"
 )
