+++ pkg/topology/kademlia/internal/metrics/metrics_test.go
@@ -9,9 +9,9 @@ import (
 	"time"
 
 	"github.com/ethersphere/bee/pkg/p2p"
+	"github.com/ethersphere/bee/pkg/shed"
 	"github.com/ethersphere/bee/pkg/swarm"
 	"github.com/ethersphere/bee/pkg/topology/kademlia/internal/metrics"
-	"github.com/ethersphere/bee/pkg/topology/kademlia/internal/shed"
 	"github.com/ethersphere/bee/pkg/util/testutil"
 	"github.com/google/go-cmp/cmp"
 )
